*: decouple FlashGrpcServerHolder from Server.cpp (#5516)
ref pingcap/tiflash#4609
