+++ dbms/src/Flash/DiagnosticsService.cpp
@@ -38,7 +38,7 @@ using diagnosticspb::SearchLogResponse;
     ::diagnosticspb::ServerInfoResponse * response)
 try
 {
-    const TiFlashRaftProxyHelper * helper = server.context().getTMTContext().getKVStore()->getProxyHelper();
+    const TiFlashRaftProxyHelper * helper = context.getTMTContext().getKVStore()->getProxyHelper();
     if (helper)
     {
         std::string req = request->SerializeAsString();
@@ -63,18 +63,18 @@ catch (const std::exception & e)
 }
 
 // get & filter(ts of last record < start-time) all files in same log directory.
-std::list<std::string> getFilesToSearch(IServer & server, Poco::Logger * log, const int64_t start_time)
+std::list<std::string> getFilesToSearch(Poco::Util::LayeredConfiguration & config, Poco::Logger * log, const int64_t start_time)
 {
     std::list<std::string> files_to_search;
 
     std::string log_dir; // log directory
-    auto error_log_file_prefix = server.config().getString("logger.errorlog", "*");
-    auto tracing_log_file_prefix = server.config().getString("logger.tracing_log", "*");
+    auto error_log_file_prefix = config.getString("logger.errorlog", "*");
+    auto tracing_log_file_prefix = config.getString("logger.tracing_log", "*");
     // ignore tiflash error log and mpp task tracing log
     std::vector<String> ignore_log_file_prefixes = {error_log_file_prefix, tracing_log_file_prefix};
 
     {
-        auto log_file_prefix = server.config().getString("logger.log");
+        auto log_file_prefix = config.getString("logger.log");
         if (auto it = log_file_prefix.rfind('/'); it != std::string::npos)
         {
             log_dir = std::string(log_file_prefix.begin(), log_file_prefix.begin() + it);
@@ -163,7 +163,7 @@ grpc::Status searchLog(Poco::Logger * log, ::grpc::ServerWriter<::diagnosticspb:
         LOG_FMT_DEBUG(log, "Handling SearchLog done: {}", request->DebugString());
     });
 
-    auto files_to_search = getFilesToSearch(server, log, start_time);
+    auto files_to_search = getFilesToSearch(config, log, start_time);
 
     for (const auto & path : files_to_search)
     {
